POI App / attempts to fix the camera external intention

+++ Android AppDev/+ Apps/PointOfView/PointOfViewApp/Poco/PointOfInterestExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using System.IO;
 using Android.Graphics;
 using Android.OS;
+using JFile = Java.IO.File;
 using Path = System.IO.Path;
-using JFile=Java.IO.File;
 
 namespace PointOfViewApp.Poco
 {
@@ -10,13 +10,15 @@ namespace PointOfViewApp.Poco
    {
       private const string AppName = "POIApp";
       private const string ImageNamePrefix = "poiimage";
+      private const string PrivateAppPrefix = "poi_images";
 
       public static string GetFileName(this PointOfInterest interest) => interest.Id.GetFileName();
 
       public static string GetFileName(this int poiId)
       {
          var storagePath = Path.Combine(Environment.ExternalStorageDirectory.Path, AppName);
-         var path = Path.Combine(storagePath, $"{ImageNamePrefix}{poiId}.jpg");
+         var path = Path.Combine(storagePath,
+            $"{PrivateAppPrefix}{Path.DirectorySeparatorChar}{ImageNamePrefix}{poiId}.jpg");
          return path;
       }
 
@@ -39,10 +41,7 @@ namespace PointOfViewApp.Poco
       public static void DeleteImage(this int poiId)
       {
          var filePath = poiId.GetFileName();
-         if (File.Exists(filePath))
-         {
-            File.Delete(filePath);
-         }
+         if (File.Exists(filePath)) File.Delete(filePath);
       }
    }
 }
